requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.29
asyncpg>=0.29.0
alembic>=1.13.1
pydantic>=2.7.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
python-multipart>=0.0.9
reportlab>=4.1.0
